+++ src/fidesops/ops/schemas/saas/saas_config.py
@@ -212,6 +212,7 @@ class SaaSType(Enum):
 
     adobe_campaign = "adobe_campaign"
     auth0 = "auth0"
+    logi_id = "logi_id"
     mailchimp = "mailchimp"
     hubspot = "hubspot"
     outreach = "outreach"