+++ tests/ops/api/v1/endpoints/test_connection_template_endpoints.py
@@ -41,7 +41,7 @@ class TestGetConnections:
         resp = api_client.get(url, headers=auth_header)
         data = resp.json()["items"]
         assert resp.status_code == 200
-        assert len(data) == 20
+        assert len(data) == 21
 
         assert {
             "identifier": ConnectionType.postgres.value,
@@ -94,7 +94,7 @@ class TestGetConnections:
         resp = api_client.get(url + "?system_type=saas", headers=auth_header)
         assert resp.status_code == 200
         data = resp.json()["items"]
-        assert len(data) == 12
+        assert len(data) == 13
 
         resp = api_client.get(url + "?system_type=database", headers=auth_header)
         assert resp.status_code == 200